# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  sds-admin_internal_dto.CreateRecordTypeRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 20
        type: string
    required:
    - name
    type: object
...
  sds-admin_internal_dto.ErrorResponse:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
...
  sds-admin_internal_dto.RecordTypeResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  sds-admin_internal_dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  sds-admin_internal_dto.UpdateRecordTypeRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 20
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: SDS Admin API for domain management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SDS Admin API
  version: "1.0"
paths:
...
  /record-types:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all record types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/sds-admin_internal_dto.RecordTypeResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: List all record types
      tags:
      - record-types
    post:
      consumes:
      - application/json
      description: Create a new record type with the provided information
      parameters:
      - description: Record type creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sds-admin_internal_dto.CreateRecordTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.RecordTypeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Create a new record type
      tags:
      - record-types
  /record-types/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a record type by its ID (soft delete)
      parameters:
      - description: Record Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Delete a record type
      tags:
      - record-types
    get:
      consumes:
      - application/json
      description: Retrieve a single record type by its ID
      parameters:
      - description: Record Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.RecordTypeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Get a record type by ID
      tags:
      - record-types
    put:
      consumes:
      - application/json
      description: Update an existing record type with new information
      parameters:
      - description: Record Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Record type update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sds-admin_internal_dto.UpdateRecordTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.RecordTypeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Update a record type
      tags:
      - record-types
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"